refactor(contracts): replace structs with common definitions
