+++ common/src/node/mod.rs
@@ -11,8 +11,9 @@ use super::*;
 
 #[derive(Debug, Clone, Eq, PartialEq, BorshDeserialize, BorshSerialize, Deserialize, Serialize)]
 pub struct HumanReadableNode {
-    pub account_id:       near_sdk::AccountId,
-    pub multi_addr:       String,
-    pub balance:          u128,
-    pub bn254_public_key: Vec<u8>,
+    pub account_id:         near_sdk::AccountId,
+    pub multi_addr:         String,
+    pub balance:            u128,
+    pub bn254_public_key:   Vec<u8>,
+    pub ed25519_public_key: Vec<u8>,
 }
