+++ contracts/Cargo.toml
@@ -19,6 +19,7 @@ near-contract-standards = { workspace = true }
 near-sdk = { workspace = true, features = ["legacy", "unit-testing"] }
 near-sys = { workspace = true }
 schemars = { workspace = true }
+seda-common = { workspace = true }
 serde = { workspace = true }
 sha2 = { workspace = true, features = ["std"] }
 uint = { workspace = true }
