+++ contracts/src/node_registry_test.rs
@@ -4,6 +4,7 @@ use near_sdk::{
     test_utils::get_logs,
     testing_env,
 };
+use seda_common::{HumanReadableNode, UpdateNode};
 
 use super::test_utils::{
     bn254_sign,
@@ -74,7 +75,9 @@ fn set_node_multi_addr() {
     assert_eq!(get_logs(), vec!["bob_near registered node"]);
 
     // update the multi_addr
-    contract.update_node(UpdateNode::SetSocketAddress("1.1.1.1:8081".to_string()));
+    contract.update_node(UpdateNode::SetSocketAddress {
+        new_multi_addr: "1.1.1.1:8081".to_string(),
+    });
 
     // check the multi_addr after updating
     testing_env!(get_context_view());
@@ -238,10 +241,7 @@ fn deposit_withdraw() {
     );
 
     // check alice is not active
-    assert_eq!(
-        contract.is_node_active("alice_near".to_string().try_into().unwrap()),
-        false
-    );
+    assert!(!contract.is_node_active("alice_near".to_string().try_into().unwrap()),);
 
     // check alice's deposited amount
     let node_balance = contract.get_node_balance("alice_near".to_string().try_into().unwrap());
